chore: divide api tests into smaller files and move them to correct folders
