+++ tests/api/smoke/articles.smoke.api.spec.ts
@@ -1,7 +1,7 @@
 import { apiEndpoints } from '@_src/api/utils/api.util';
 import { expect, test } from '@_src/ui/fixtures/merge.fixture';
 
-test.describe('Verify articles API endpoint @GAD-R08-01 @smoke', () => {
+test.describe('Verify articles API endpoint @GAD-R08-01 @smoke @articles', () => {
   test.describe('Verify each condition in separate test', () => {
     test('GET articles should return status code 200', async ({ request }) => {
       // Arrange
